cleaning up ts errors
