--- build-sd.js
@@ -1,2 +1,0 @@
-require = require("esm")(module);
-module.exports = require("./sd.config.mjs").default;
