+++ opacity/doc/index.mdx
@@ -1,9 +1,9 @@
 import { mdx } from "@mdx-js/react";
 import { Layout } from "~/layout";
-import OpacityCard from '../src/OpacityCard';
+import OpacityShowcase from "../src/OpacityShowcase";
 import theme from "~/theme/src/theme.json";
 export default Layout;
 
 ## Opacity
 
-<OpacityCard theme={theme} />
+<OpacityShowcase theme={theme} />
