+++ package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@backlight-dev/graylog.sawmill-sd",
-  "version": "0.0.0",
+  "name": "@graylog/sawmill",
+  "version": "0.0.1",
   "scripts": {
     "serve": "npx backlight@latest serve f8bLRusG6EujzpCzDM6X --open",
-    "build": "node build-sd.js"
+    "build": "node theme/build.js"
   },
   "dependencies": {
     "@divriots/dockit-react": "0.0.72",
@@ -19,6 +19,8 @@
     "styled-components": "^5.3.1"
   },
   "devDependencies": {
-    "@types/react": "^17.0.35"
+    "@types/mdx-js__react": "^1.5.5",
+    "@types/react": "^17.0.35",
+    "@types/styled-components": "^5.1.15"
   }
 }
