+++ radii/src/RadiiShowcase.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import styled from "styled-components";
 
 const Box = styled.div`
