+++ theme/build.js
@@ -1,0 +1,4 @@
+// Styled Dictionary Node build script
+// Run with `npm run build`
+require = require("esm")(module);
+module.exports = require("../sd.config.mjs").default;
