+++ theme/src/variables.css
@@ -1,6 +1,6 @@
 /**
  * Do not edit directly
- * Generated on Tue, 23 Nov 2021 16:17:31 GMT
+ * Generated on Tue, 23 Nov 2021 17:51:37 GMT
  */
 
 :root {
